feat: Working login -> home page
