+++ frontend/src/HomePage.jsx
@@ -1,0 +1,13 @@
+// src/Dashboard.jsx
+import React from "react";
+
+function Dashboard() {
+    return (
+        <div>
+            <h1>Testing</h1>
+            <p>Testing</p>
+        </div>
+    );
+}
+
+export default Dashboard;
